Mutation operators

Mutation operators (a.k.a. mutators) are derived from the EvoLP.MutationMethod abstract type, and some of them have parameters to control how the mutation is performed. Mutation methods are dependent on both the data contained in an individual and its representation.

Currently, mutation is only implemented for vector individuals.

Selecting a mutation operator

EvoLP provides many built-in mutators.

For binary vectors

EvoLP.BitwiseMutation — Type

Bitwise mutation with probability λ of flipping each bit.

For continuous vectors

EvoLP.GaussianMutation — Type

Gaussian mutation with standard deviation σ, which should be a real number.

For permutation-based vectors

EvoLP.SwapMutation — Type

Swap mutation for permutation-based individuals.

EvoLP.InsertMutation — Type

Insert mutation for permutation-based individuals.

EvoLP.ScrambleMutation — Type

Scramble mutation for permutation-based individuals.

EvoLP.InversionMutation — Type

Inversion mutation for permutation-based individuals.

Performing the mutation

After "instantiating" a mutation method, you can use mutate on a single individual ind. All operators return a copy, and in the process no individual is modified.

EvoLP.mutate — Function

mutate(M::BitwiseMutation, ind)

Randomly flips each bit with a probability λ.

mutate(M::GaussianMutation, ind)

Randomly add Gaussian noise to the ind candidate solution, with a standard deviation of σ.

mutate(::SwapMutation, ind)

Randomly swap the position of two alleles in the ind candidate solution.

mutate(::InsertMutation, ind)

Randomly choose two positions a and b from ind, insert at a+1 the element at position b`, and shift the rest of the elements.

mutate(::ScrambleMutation, ind)

Randomly scramble the subsequence between two random points in ind.

mutate(::InversionMutation, ind)

Invert the subsequence between two random points in ind.