Selection operators

Parent selection operators (a.k.a. selectors) in EvoLP are based on fitness and are used to select individuals for crossover. The selectors always return indices so that individuals can be selected from the population later.

These methods come in two variants: steady-state and generational. Steady-state operators return two parent indices, while generational operators return a list of two parent indices for each fitness in the population.

In other words, steady-state variants perform one selection, while generational variants perform $n$ selections.

Regardless of the variant, all operators are derived from the EvoLP.SelectionMethod abstract type and some of them have parameters you can adjust.

Choosing a selection operator

EvoLP provides many built-in selectors.

EvoLP.TournamentSelectionSteady — Type

Tournament parent selection with tournament size k.

EvoLP.TournamentSelectionGenerational — Type

Tournament parent selection with tournament size k.

EvoLP.TruncationSelectionSteady — Type

Truncation selection for selecting top k possible parents in the population.

EvoLP.TruncationSelectionGenerational — Type

Truncation selection for selecting top k possible parents in the population.

EvoLP.RouletteWheelSelectionSteady — Type

Roulette wheel parent selection.

EvoLP.RouletteWheelSelectionGenerational — Type

Roulette wheel parent selection.

EvoLP.RankBasedSelectionSteady — Type

Rank-based parent selection.

EvoLP.RankBasedSelectionGenerational — Type

Rank-based parent selection.

Performing the selection

After "instantiating" a selection method, you can use the select function on an array of fitnesses y to obtain parents' indices (that you will need to slice from the population in your algorithm later.)

EvoLP.select — Function

select(t::TruncationSelectionSteady, y)

Select two random parents out from the top t.k in the population.

select(t::TournamentSelectionSteady, y)

Select two parents which are the winners from two random tournaments of size t.k.

select(::RouletteWheelSelectionSteady, y)

Select two random parents with probability proportional to their fitness.

select(::RankBasedSelectionSteady, y)

Select two random parents with probability proportional to their ranks.

select(t::TruncationSelectionGenerational, y)

Select two random parents (from the top t.k) in the population for each fitness in y.

select(t::TournamentSelectionGenerational, y)

Select two parents which are the winners from two random tournaments of size t.k for each fitness in y.

select(::RouletteWheelSelectionGenerational, y)

Select two random parents with probability proportional to their fitness, for each fitness in y.

select(::RankBasedSelectionGenerational, y)

Select two random parents with probability proportional to their ranks, for each fitness in y.