Generational GA for continuous optimisation

This tutorial details how to use the built-in Genetic Algorithm (GA) on a continuous test function.

We start by importing EvoLP. We will compute some statistics using the Logbook so we need some additional modules as well:

using Statistics
using EvoLP
using OrderedCollections

For this example we will use the Rosenbrock function, which is already included as a benchmark function in EvoLP. We can look at the documentation like so:

EvoLP.rosenbrock — Function

rosenbrock(x; b=100)

Changed since EvoLP 1.3

This is the $d$-dimensional Rosenbrock function. In previous releases, it was 2d only and had an additional keyword argument a.

Update your workflow accordingly.

The $d$-dimensional Rosenbrock banana benchmark function. With $b=100$, minimum is at $f([1, \dots, 1]) = 0$

\[f(x) = \sum_{i=1}^{d-1} \left[b(x_{i+1} - x_i^2)^2 + (x_i - 1)^2 \right]\]

Implementing the solution

Let's start creating the population. We can use the normal_rand_vector_pop generator, which uses a normal distribution for initialisation:

EvoLP.normal_rand_vector_pop — Function

normal_rand_vector_pop(μ, m, C; rng=Random.GLOBAL_RNG)

Generate a population of μ vector individuals using a normal distribution with means m and covariance C.

m expects a vector of length l (i.e. length of an individual) while C expects an l x l matrix of covariances.

Examples

julia> normal_rand_vector_pop(3, [0, 0], [1 0; 0 1])
3-element Vector{Vector{Float64}}:
 [-0.15290525182234904, 0.8715880371871617]
 [-1.1283800329864322, -0.9256584563613383]
 [-0.5384758126777555, -0.8141702145510666]

pop_size = 50
population = normal_rand_vector_pop(pop_size, [0, 0], [1 0; 0 1])
first(population, 3)

3-element Vector{Vector{Float64}}:
 [-1.4282160292298753, 1.2318788291908964]
 [0.6516138627719154, 0.2116230303640737]
 [1.7260062473358426, 1.761964931940205]

In a GA, we have selection, crossover and mutation.

We can easily set up these operators using the built-ins provided by EvoLP. Let's use rank based selection and interpolation crossover with 0.5 as the scaling factor:

EvoLP.InterpolationRecombinator — Type

Interpolation crossover with scaling parameter α.

S = RankBasedSelector()
C = InterpolationRecombinator(0.5)

InterpolationRecombinator(0.5)

For mutation, we can use Gaussian noise:

EvoLP.GaussianMutator — Type

Gaussian mutation with standard deviation σ, which must be a real number.

M = GaussianMutator(0.05)

GaussianMutator(0.05)

Now we can set up the Logbook to record statistics about our run:

statnames = ["mean_eval", "max_f", "min_f", "median_f"]
fns = [mean, maximum, minimum, median]
thedict = LittleDict(statnames, fns)
thelogger = Logbook(thedict)

Logbook(OrderedCollections.LittleDict{AbstractString, Function, Vector{AbstractString}, Vector{Function}}("mean_eval" => Statistics.mean, "max_f" => maximum, "min_f" => minimum, "median_f" => Statistics.median), NamedTuple{(:mean_eval, :max_f, :min_f, :median_f)}[])

And now we're ready to use the GA built-in algorithm:

EvoLP.GA — Function

GA(f, pop, k_max, S, C, M)
GA(logbook::Logbook, f, population, k_max, S, C, M)
GA(notebooks::Vector{Logbook}, f, population, k_max, S, C, M)

Generational Genetic Algorithm.

Arguments

f::Function: objective function to minimise.

population::AbstractVector: a list of vector individuals.

k_max::Integer: number of iterations.

S::ParentSelector: one of the available ParentSelector.

C::Recombinator: one of the available Recombinator.

M::Mutator: one of the available Mutator.

Returns a Result.

result = GA(thelogger, rosenbrock, population, 300, S, C, M);

Result(0.05219730421398745, [0.8053522637095605, 0.4665194617737337], [[0.7097359075031934, 0.47280146179303295], [0.6962525109810143, 0.6394872716702015], [0.7298183116156254, 0.6189728814525586], [0.6160097365826285, 0.5301503727909843], [0.8515882322692199, 0.5598662894181504], [0.6760605744129596, 0.4878450750488618], [0.6611762893874242, 0.524368333452694], [0.6763207353379527, 0.5020378499588202], [0.7235843794801864, 0.5564603794513797], [0.7942725826373946, 0.6072178211084173]  …  [0.6821342587305803, 0.5373834933786094], [0.7638707486573149, 0.5368922732525021], [0.7491219812139274, 0.5549463787236234], [0.7020332637710669, 0.5088372467799261], [0.7069003524325861, 0.5314321487186541], [0.7888906960855152, 0.46847636961482453], [0.7427345816871863, 0.6189063656045113], [0.8044562373230385, 0.6064634050094748], [0.6445805814932365, 0.5020297951320906], [0.7125004278899644, 0.516110293457282]], 300, 15000, 1.43552039)

The output was suppressed so that we can analyse each part of the result separately using functions instead:

@show optimum(result)

0.05219730421398745

@show optimizer(result)

2-element Vector{Float64}:
 0.8053522637095605
 0.4665194617737337

@show f_calls(result)

15000

thelogger.records[end]

(mean_eval = 1.0467740970549924, max_f = 4.03563698963417, min_f = 0.05219730421398745, median_f = 0.6643835760466139)

The records in the Logbook are NamedTuples. This makes it easier to export and analyse using DataFrames, for example:

using DataFrames
DataFrame(thelogger.records)

500×4 DataFrame
 Row │ mean_eval   max_f       min_f        median_f  
     │ Float64     Float64     Float64      Float64   
─────┼────────────────────────────────────────────────
   1 │ 22.0251     406.9       0.447041     6.78992
   2 │  3.61617     36.062     0.124031     1.96466
   3 │  1.13189      3.18343   0.127583     1.07601
   4 │  0.781777     1.6644    0.309661     0.711803
   5 │  0.593735     0.935043  0.294026     0.588684
   6 │  0.527621     0.766033  0.315916     0.518089
   7 │  0.522381     0.745129  0.37027      0.527158
   8 │  0.493569     0.807639  0.275269     0.498038
  ⋮ │     ⋮           ⋮            ⋮           ⋮