Using PSO to minimise the Michalewicz function

This tutorial showcases how to use the built-in Particle Swarm Optimisation (PSO) algorithm for finding the minimum in a continuos setting.

We start by importing our necessary modules

using EvoLP
using Statistics
using OrderedCollections

For this example, we will use the Michalewicz function, which is a test function included in EvoLP:

EvoLP.michalewicz — Function

michalewicz(x; m=10)

The Michalewicz function is a $d$-dimensional function with several steep valleys, where m controls the steepness. m is usually set at 10. For 2 dimensions, $x^* = [2.20, 1.57]$, with $f(x^*) = -1.8011$.

\[f(x) = -\sum_{i=1}^{d}\sin(x_i) \sin^{2m}\left(\frac{ix_i^2}{\pi}\right)\]

In this case we will use d=2 and m=10, which are the default values implemented.

In PSO, we use particles. Each particle has a position and a velocity, and remembers the best position the whole swarm has visited. We can create a population of particles in multiple ways, but EvoLP provides 2 particle generators with random positions: either uniform or following a normal distribution.

Let's use the normal generator:particle generators

EvoLP.normal_rand_particle_pop — Function

normal_rand_particle_pop(μ, m, C; y=Inf, rng=Random.GLOBAL_RNG)

Generate a population of μ Particle using a normal distribution with means mand covarianceC`.

m expects a vector of length l (i.e. number of dimensions) while C expects an l x l matrix of covariances.

y is the evaluation and current best value. y is set to Inf by default.

Examples

julia> normal_rand_particle_pop(3, [0, 0], [1 0; 0 1])
3-element Vector{Particle}:
 Particle([-0.6025996585348097, -1.0055548956861133], [0.0, 0.0], Inf, [-0.6025996585348097, -1.0055548956861133], Inf)
 Particle([-0.7562454555135321, 1.9490439959687778], [0.0, 0.0], Inf, [-0.7562454555135321, 1.9490439959687778], Inf)
 Particle([0.5687241357408321, -0.7406267072113427], [0.0, 0.0], Inf, [0.5687241357408321, -0.7406267072113427], Inf)

Since we are using the 2-dimensional version of the function, we need to provide a vector of 2 means and a 2 \times 2 matrix of covariances:

population = normal_rand_particle_pop(50, [0, 0], [1 0; 0 1])
first(population, 3)

3-element Vector{Particle}:
 Particle([0.7839927550634601, 0.6804724692646161], [0.0, 0.0], Inf, [0.7839927550634601, 0.6804724692646161], Inf)
 Particle([-0.4071111189054881, -1.6921494385950955], [0.0, 0.0], Inf, [-0.4071111189054881, -1.6921494385950955], Inf)
 Particle([0.7312702202363397, 0.7976019813479507], [0.0, 0.0], Inf, [0.7312702202363397, 0.7976019813479507], Inf)

We can use the Logbook to save information about each iteration of the run. Let's save the average, median and best fitness:

statnames = ["avg_fit", "median_fit", "best_fit"]
callables = [mean, median, minimum]
thedict = LittleDict(statnames, callables)
logbook = Logbook(thedict)

Logbook(OrderedCollections.LittleDict{AbstractString, Function, Vector{AbstractString}, Vector{Function}}("avg_fit" => Statistics.mean, "median_fit" => Statistics.median, "best_fit" => minimum), NamedTuple{(:avg_fit, :median_fit, :best_fit)}[])

Now we can use the built-in PSO algorithm:

EvoLP.PSO — Function

PSO(f, population, k_max; w=1, c1=1, c2=1)
PSO(logger::Logbook, f, population, k_max; w=1, c1=1, c2=1)

Arguments

f::Function: Objective function to minimise.

population::Vector{Particle}: a list of Particle individuals.

k_max::Integer: number of iterations.

Keywords

w: inertia weight. Optional, by default 1.

c1: cognitive coefficient (own's position). Optional, by default 1.

c2: social coefficient (others' position). Optional, by default 1.

Returns a Result.

Let's use the default parameters, and 30 iterations:

result = PSO(logbook, michalewicz, population, 30);

Result(-1.8008301892001088, [2.204710831252206, 1.5675716169753269], Particle[Particle([2.095989976564521, 1.5406847541151736], [0.05441677294982039, 0.03939097583317995], -1.6063872073002252, [2.2130955010336915, 1.5629841778246916], -1.7971533760251521), Particle([2.011527119404721, 2.46589199246142], [0.3647161788301434, -2.5011963902903087], -0.40218422821557565, [1.675040245817072, 1.5563477654021183], -0.9983948541874456), Particle([2.0177075108695113, 1.3291786533339442], [0.526723127848443, 0.584205109982869], -0.5432453196317193, [2.3683907743347334, 1.6122247713938282], -1.369998196081324), Particle([2.2937336292154984, 1.567824770689598], [-0.32615796942781483, -0.0013477414474710183], -1.6726442428905561, [2.1648795383814363, 1.567013729948652], -1.7779315720657651), Particle([2.010861859458455, 3.3386537427319176], [-0.13602082273856253, 0.7526122501398964], -0.3998072646705076, [2.0298862009631744, 1.6613613699500211], -1.156628141076439), Particle([2.1065139547196368, 1.638538498736107], [0.45105423315808785, -0.08555671620215863], -1.4918032413408653, [2.201658873032149, 1.57839683885308], -1.798930046770035), Particle([2.1976782271670867, 1.566974415700208], [0.008881651620177415, 0.0004522301796888942], -1.800270442371624, [2.2008810388229545, 1.5667306876652494], -1.8005693431074064), Particle([2.5624106956953008, 2.524328015972313], [1.2143107063865561, 0.03169373747587778], -0.03796086469255151, [2.1432543237579034, 1.5385241528095002], -1.7063264607328872), Particle([2.2010917521085362, 1.5804791058559386], [-0.0069133653508024855, 0.0020197285323280013], -1.7974367574497707, [2.204710831252206, 1.5675716169753269], -1.8008301892001088), Particle([3.510884756082791, -0.2071713218322211], [1.567315210228335, 1.4057479898160534], 0.000329220280211798, [1.9161530110268954, 1.4781422978867327], -0.8978104762815577)  …  Particle([0.4157306935869438, -0.89923447789785], [5.421916677969315, 1.4601076916544877], 5.484287980402368e-7, [2.308572307570737, 1.1256148103036243], -0.6329868015422921), Particle([2.6385569121082106, 1.1042591392403178], [-0.9576655117876909, 0.23762129918436517], -0.006137847972420826, [2.024356908641573, 1.5230365311205412], -1.352563668866258), Particle([2.1768359798448342, 1.513354260796507], [0.14103893759225283, 0.08347186164457915], -1.6693206208312885, [2.150043538062189, 1.5627793539661106], -1.7555086057262534), Particle([3.321086439141054, 1.5839253479059108], [0.10376732980757353, -0.01433397714555198], -0.9929851290508247, [2.1201293855696646, 1.583504795726538], -1.6940293054442956), Particle([2.4142621952891345, 1.1261289468147748], [-0.5270679284426713, 0.8177073312721539], -0.2939810875539917, [2.236549800585813, 1.5384982264427631], -1.7422683607176084), Particle([2.069587979999393, 7.772788021745232], [-0.7395349017108649, -4.277726553403339], -0.5699731993458337, [2.121121179380463, 1.572418396242409], -1.702703889504262), Particle([3.458556255890325, 2.8189516935286836], [-0.7718642653011618, -0.5428057640925146], -0.09307861072079268, [1.9731497813497119, 1.5958275234794406], -1.274838913066102), Particle([0.12593824069225867, 1.0331542310342343], [-0.2961622471449722, 0.7375696836169952], -7.926581589703381e-5, [2.116352059353244, 1.575374104485069], -1.6911726464063466), Particle([4.498401696466321, 1.9872865072402508], [-1.596562671088563, 0.9179136600392725], -2.1597156386986362e-5, [3.002327458145496, 1.5630748434205548], -0.9975998366611393), Particle([5.03178950629939, 8.401452101616055], [-1.779779022974766, -13.006322992747835], 0.6067373177818212, [1.3479204691710998, 1.5737169286176917], -0.9996594950014577)], 30, 1550, 0.31558895111083984)

The output was suppressed so that we can analyse each part of the result separately using the Result functions:

@show optimum(result)

-1.8008301892001088

@show optimizer(result)

2-element Vector{Float64}:
 2.204710831252206
 1.5675716169753269

@show f_calls(result)

1550

We can also take a look at the logbook's records and see how the calculated statistics changed throughout the run:

for (i, I) in enumerate(logbook.records)
    print("it: $(i) with best_pos: $(I[3]) and avg_pos: $(I[1]) \n")
end

it: 1 with best_pos: -0.8083097634277963 and avg_pos: 0.007109715208891737
it: 2 with best_pos: -0.9996594950014577 and avg_pos: -0.17070646802650494
it: 3 with best_pos: -0.9977243019214357 and avg_pos: -0.2048542635529755
it: 4 with best_pos: -1.0026922523038357 and avg_pos: -0.2747951995156524
it: 5 with best_pos: -1.5510091316370413 and avg_pos: -0.45988750114090166
it: 6 with best_pos: -1.7382821061146956 and avg_pos: -0.5843718797458269
it: 7 with best_pos: -1.7686240265867825 and avg_pos: -0.5214602192147078
it: 8 with best_pos: -1.7833432538377956 and avg_pos: -0.6581402577467281
it: 9 with best_pos: -1.793524624672582 and avg_pos: -0.6804905635770132
it: 10 with best_pos: -1.79316196350381 and avg_pos: -0.6044092540947791
it: 11 with best_pos: -1.7989585082784565 and avg_pos: -0.7485541904747328
it: 12 with best_pos: -1.7995777639691042 and avg_pos: -0.5760511942993649
it: 13 with best_pos: -1.7975190921631758 and avg_pos: -0.5989305890740995
it: 14 with best_pos: -1.8008301892001088 and avg_pos: -0.6708532724091174
it: 15 with best_pos: -1.7971533760251521 and avg_pos: -0.6152464322279342
it: 16 with best_pos: -1.786066417447012 and avg_pos: -0.6230472440525985
it: 17 with best_pos: -1.800793926455185 and avg_pos: -0.6745567657887307
it: 18 with best_pos: -1.798930046770035 and avg_pos: -0.6582068220196572
it: 19 with best_pos: -1.7900231242059592 and avg_pos: -0.5998065500856963
it: 20 with best_pos: -1.7938168995070591 and avg_pos: -0.6643313434536949
it: 21 with best_pos: -1.7999313677868845 and avg_pos: -0.542939375225023
it: 22 with best_pos: -1.7816436780374354 and avg_pos: -0.6406444420479456
it: 23 with best_pos: -1.7965934655030356 and avg_pos: -0.6244526513106952
it: 24 with best_pos: -1.7977250286715356 and avg_pos: -0.5803149052974892
it: 25 with best_pos: -1.790270501438703 and avg_pos: -0.6255101036629112
it: 26 with best_pos: -1.7950818706934033 and avg_pos: -0.516712194569448
it: 27 with best_pos: -1.8004958665546158 and avg_pos: -0.5757563719248763
it: 28 with best_pos: -1.8005693431074064 and avg_pos: -0.5627105362064377
it: 29 with best_pos: -1.799939359576145 and avg_pos: -0.5724932115051841
it: 30 with best_pos: -1.800270442371624 and avg_pos: -0.5792356717267855