Using PSO to minimise the Michalewicz function

This tutorial showcases how to use the built-in Particle Swarm Optimisation (PSO) algorithm for finding the minimum in a continuos setting.

We start by importing our necessary modules

using EvoLP
using Statistics
using OrderedCollections

For this example, we will use the Michalewicz function, which is a test function included in EvoLP:

EvoLP.michalewicz — Function

michalewicz(x; m=10)

The Michalewicz function is a $d$-dimensional function with several steep valleys, where m controls the steepness. m is usually set at 10. For 2 dimensions, $x^* = [2.20, 1.57]$, with $f(x^*) = -1.8011$.

\[f(x) = -\sum_{i=1}^{d}\sin(x_i) \sin^{2m}\left(\frac{ix_i^2}{\pi}\right)\]

In this case we will use d=2 and m=10, which are the default values implemented.

In PSO, we use particles. Each particle has a position and a velocity, and remembers the best position the whole swarm has visited. We can create a population of particles in multiple ways, but EvoLP provides 2 particle generators with random positions: either uniform or following a normal distribution.

Let's use the normal generator:particle generators

EvoLP.normal_rand_particle_pop — Function

normal_rand_particle_pop(n, μ, Σ; y=Inf, rng=Random.GLOBAL_RNG)

Generate a population of n Particle using a normal distribution with means μand covarianceΣ`.

μ expects a vector of length l (i.e. number of dimensions) while Σ expects an l x l matrix of covariances.

y is the evaluation and current best value. y is set to Inf by default.

Examples

julia> normal_rand_particle_pop(3, [0, 0], [1 0; 0 1])
3-element Vector{Particle}:
 Particle([-0.6025996585348097, -1.0055548956861133], [0.0, 0.0], Inf, [-0.6025996585348097, -1.0055548956861133], Inf)
 Particle([-0.7562454555135321, 1.9490439959687778], [0.0, 0.0], Inf, [-0.7562454555135321, 1.9490439959687778], Inf)
 Particle([0.5687241357408321, -0.7406267072113427], [0.0, 0.0], Inf, [0.5687241357408321, -0.7406267072113427], Inf)

Since we are using the 2-dimensional version of the function, we need to provide a vector of 2 means and a 2 \times 2 matrix of covariances:

population = normal_rand_particle_pop(50, [0, 0], [1 0; 0 1])
first(population, 3)

3-element Vector{Particle}:
 Particle([1.4283020328561498, -0.033090934612689235], [0.0, 0.0], Inf, [1.4283020328561498, -0.033090934612689235], Inf)
 Particle([-0.8940186599732496, 0.02339875908371508], [0.0, 0.0], Inf, [-0.8940186599732496, 0.02339875908371508], Inf)
 Particle([-1.4466933970400233, 0.20574079019192296], [0.0, 0.0], Inf, [-1.4466933970400233, 0.20574079019192296], Inf)

We can use the Logbook to save information about each iteration of the run. Let's save the average, median and best fitness:

statnames = ["avg_fit", "median_fit", "best_fit"]
callables = [mean, median, minimum]
thedict = LittleDict(statnames, callables)
logbook = Logbook(thedict)

Logbook(OrderedCollections.LittleDict{AbstractString, Function, Vector{AbstractString}, Vector{Function}}("avg_fit" => Statistics.mean, "median_fit" => Statistics.median, "best_fit" => minimum), NamedTuple{(:avg_fit, :median_fit, :best_fit)}[])

Now we can use the built-in PSO algorithm:

EvoLP.PSO — Function

PSO(f, population, k_max; w=1, c1=1, c2=1)
PSO(logger::Logbook, f, population, k_max; w=1, c1=1, c2=1)

Arguments

f::Function: Objective function to minimise.

population::Vector{Particle}: a list of Particle individuals.

k_max::Integer: number of iterations.

Keywords

w: inertia weight. Optional, by default 1.

c1: cognitive coefficient (own's position). Optional, by default 1.

c2: social coefficient (others' position). Optional, by default 1.

Returns a Result.

Let's use the default parameters, and 30 iterations:

result = PSO(logbook, michalewicz, population, 30);

Result(-1.8007981001657498, [2.197572075384775, 1.5718405495977332], Particle[Particle([1.1555358154888706, 1.5217941579583245], [0.3824515725656024, -0.16168127682177516], -0.9098992721945978, [2.191177341727804, 1.5734127291656823], -1.7988045518711044), Particle([2.6864512209850018, 2.1418322138033985], [0.14263009146084732, -0.9682840893402276], -0.0013052162265253843, [2.284067630794027, 1.4761685479131084], -1.4069553257654062), Particle([2.0058437827207936, 1.6543615651856627], [3.7769282957557375, -0.052816114012111975], -1.1274853099194997, [2.010833229225755, 1.5711921843389598], -1.400047730133554), Particle([2.3226695124305063, 1.598253032811977], [-4.047774053898121, -0.21223354055909033], -1.5584144156784743, [2.3226695124305063, 1.598253032811977], -1.5584144156784743), Particle([-5.0134772016028215, -0.09149936375399204], [3.304760235112081, 0.9025252245608011], -0.7694486929368445, [2.4640703966188466, 1.580720648244677], -1.1602834058444063), Particle([2.3513867182681083, 1.6520079360367046], [1.56884221913719, 0.15747476364569932], -1.2498455050993946, [2.0603432379039766, 1.6563850260240276], -1.2730010798477243), Particle([-4.215556232015077, 0.465259811153661], [1.6032743995544165, 2.340398270581596], -2.0274912419703383e-5, [1.043497663261122, 1.540782168358455], -0.9648138562328747), Particle([2.539216441550658, 1.7144649514882961], [-0.11426087231720068, 0.22781033540641402], -0.44634567371002953, [2.314554769846104, 1.6221302361263985], -1.5093846365874501), Particle([2.360606315815455, 1.9416135187886345], [-0.3890792014335641, -0.4605870155384265], -0.4653042529375791, [2.319013392738055, 1.6356573038398963], -1.4373478189866853), Particle([2.197572075384775, 1.5718405495977332], [-0.027166242267272475, 0.001981530285014971], -1.8007981001657498, [2.197572075384775, 1.5718405495977332], -1.8007981001657498)  …  Particle([9.748439441634773, 1.5775534449333657], [-6.849595110371294, 0.1278249458088329], -0.9390490985138895, [1.356458605544597, 1.579693280389865], -0.9967881851881255), Particle([2.7107794215610093, 1.7100082781729875], [-0.01083264600657216, 0.5084812737186087], -0.42064201553253583, [1.4748600882676701, 1.4887800696326932], -0.7714386075393337), Particle([4.156951228084455, 1.5740886698541359], [-1.21354805782175, 0.30317875869981276], -0.9987739275094825, [2.2126291091130774, 1.4324944072371668], -1.2879206829974041), Particle([3.155902880952693, 1.9790018795070499], [-0.27857087254570734, -0.49870992662402425], -3.811861610387191e-5, [2.105883519015006, 1.5863861761133597], -1.657084522911271), Particle([5.704885454593252, 0.723474642535316], [-2.5079181814389537, -1.1381839861821466], 0.007051357094016687, [0.45199852448757905, 1.580130258047187], -0.996456895583612), Particle([3.583376098446766, -0.4540116199801345], [-0.3243655743144449, 0.9040587195831676], 0.006461884201402592, [2.475707092643561, 1.5416640139197946], -1.1072176571311085), Particle([5.496887519439682, 1.5783115151502263], [-6.973358644932505, 0.018712436734526876], -0.9977043702946291, [2.515371848538586, 1.5324796577619075], -1.0203858685933735), Particle([3.94888771236684, 2.267108884287039], [-1.28975619229225, -0.22963839921786516], 0.381691802441356, [2.141704977094848, 1.4394461816279265], -1.2666440243833823), Particle([1.6957756343633665, 1.8789891135004755], [0.8318401504665557, -0.2729411100141135], -0.016085944543033234, [2.220827902980075, 1.5459313977210043], -1.7717005540318438), Particle([1.8857310554315434, 1.7608521111213462], [-0.11062230877885049, -0.5954710014698436], -0.3144588066659634, [2.086831730428917, 1.5684720508014682], -1.6171840009848246)], 30, 1550, 0.3843879699707031)

The output was suppressed so that we can analyse each part of the result separately using the Result functions:

@show optimum(result)

-1.8007981001657498

@show optimizer(result)

2-element Vector{Float64}:
 2.197572075384775
 1.5718405495977332

@show f_calls(result)

1550

We can also take a look at the logbook's records and see how the calculated statistics changed throughout the run:

for (i, I) in enumerate(logbook.records)
    print("it: $(i) with best_pos: $(I[3]) and avg_pos: $(I[1]) \n")
end

it: 1 with best_pos: -0.8657580035876815 and avg_pos: -0.0012170385447373832
it: 2 with best_pos: -0.8657580035876815 and avg_pos: -0.10032108128806168
it: 3 with best_pos: -0.9812808253949845 and avg_pos: -0.31551243818614305
it: 4 with best_pos: -0.9968813852769467 and avg_pos: -0.2619652233022636
it: 5 with best_pos: -0.999999511226863 and avg_pos: -0.32605660891678867
it: 6 with best_pos: -0.9996938074580445 and avg_pos: -0.4469192824887817
it: 7 with best_pos: -0.999853185667781 and avg_pos: -0.45896806196052325
it: 8 with best_pos: -0.9994804941466119 and avg_pos: -0.44146050257503
it: 9 with best_pos: -0.9999773270847023 and avg_pos: -0.4414862269235067
it: 10 with best_pos: -1.5555148395139695 and avg_pos: -0.5021996809196607
it: 11 with best_pos: -1.2158228913900495 and avg_pos: -0.4470248441601987
it: 12 with best_pos: -1.7010654417229962 and avg_pos: -0.561679175960725
it: 13 with best_pos: -1.6823821273468913 and avg_pos: -0.5921701923640997
it: 14 with best_pos: -1.679343231117683 and avg_pos: -0.5492863281512291
it: 15 with best_pos: -1.6181651200380354 and avg_pos: -0.5064428814054429
it: 16 with best_pos: -1.7909118597417812 and avg_pos: -0.4559379550085608
it: 17 with best_pos: -1.7285753057683393 and avg_pos: -0.5412016269524226
it: 18 with best_pos: -1.7051504622503793 and avg_pos: -0.3591388337537698
it: 19 with best_pos: -1.7567776735795713 and avg_pos: -0.4456397346254859
it: 20 with best_pos: -1.7988045518711044 and avg_pos: -0.5318327382966386
it: 21 with best_pos: -1.7778695397483095 and avg_pos: -0.44891911116340105
it: 22 with best_pos: -1.798834279785546 and avg_pos: -0.5534332387121953
it: 23 with best_pos: -1.8002653823006214 and avg_pos: -0.5027027525538091
it: 24 with best_pos: -1.798379440827754 and avg_pos: -0.4474566178956105
it: 25 with best_pos: -1.7867420596150747 and avg_pos: -0.46971026101481894
it: 26 with best_pos: -1.79603869882738 and avg_pos: -0.432703142217719
it: 27 with best_pos: -1.7929540710872383 and avg_pos: -0.4586668827710848
it: 28 with best_pos: -1.7796384514528212 and avg_pos: -0.48614100448767067
it: 29 with best_pos: -1.7894151943909111 and avg_pos: -0.41870516177669415
it: 30 with best_pos: -1.8007981001657498 and avg_pos: -0.43120883491672385