Using PSO to minimise the Michalewicz function

This tutorial showcases how to use the built-in Particle Swarm Optimisation (PSO) algorithm for finding the minimum in a continuos setting.

We start by importing our necessary modules

using EvoLP
using Statistics
using OrderedCollections

For this example, we will use the Michalewicz function, which is a test function included in EvoLP:

EvoLP.michalewicz — Function

michalewicz(x; m=10)

The Michalewicz function is a $d$-dimensional function with several steep valleys, where m controls the steepness. m is usually set at 10. For 2 dimensions, $x^* = [2.20, 1.57]$, with $f(x^*) = -1.8011$.

\[f(x) = -\sum_{i=1}^{d}\sin(x_i) \sin^{2m}\left(\frac{ix_i^2}{\pi}\right)\]

In this case we will use d=2 and m=10, which are the default values implemented.

In PSO, we use particles. Each particle has a position and a velocity, and remembers the best position the whole swarm has visited. We can create a population of particles in multiple ways, but EvoLP provides 2 particle generators with random positions: either uniform or following a normal distribution.

Let's use the normal generator:particle generators

EvoLP.normal_rand_particle_pop — Function

normal_rand_particle_pop(μ, m, C; y=Inf, rng=Random.GLOBAL_RNG)

Generate a population of μ Particle using a normal distribution with means mand covarianceC`.

m expects a vector of length l (i.e. number of dimensions) while C expects an l x l matrix of covariances.

y is the evaluation and current best value. y is set to Inf by default.

Examples

julia> normal_rand_particle_pop(3, [0, 0], [1 0; 0 1])
3-element Vector{Particle}:
 Particle([-0.6025996585348097, -1.0055548956861133], [0.0, 0.0], Inf, [-0.6025996585348097, -1.0055548956861133], Inf)
 Particle([-0.7562454555135321, 1.9490439959687778], [0.0, 0.0], Inf, [-0.7562454555135321, 1.9490439959687778], Inf)
 Particle([0.5687241357408321, -0.7406267072113427], [0.0, 0.0], Inf, [0.5687241357408321, -0.7406267072113427], Inf)

Since we are using the 2-dimensional version of the function, we need to provide a vector of 2 means and a 2 \times 2 matrix of covariances:

population = normal_rand_particle_pop(50, [0, 0], [1 0; 0 1])
first(population, 3)

3-element Vector{Particle}:
 Particle([1.0561697587202845, 0.21630944572959176], [0.0, 0.0], Inf, [1.0561697587202845, 0.21630944572959176], Inf)
 Particle([-2.143409734241095, 0.9958450352235881], [0.0, 0.0], Inf, [-2.143409734241095, 0.9958450352235881], Inf)
 Particle([0.8175180787417627, 1.599738123344914], [0.0, 0.0], Inf, [0.8175180787417627, 1.599738123344914], Inf)

We can use the Logbook to save information about each iteration of the run. Let's save the average, median and best fitness:

statnames = ["avg_fit", "median_fit", "best_fit"]
callables = [mean, median, minimum]
thedict = LittleDict(statnames, callables)
logbook = Logbook(thedict)

Logbook(OrderedCollections.LittleDict{AbstractString, Function, Vector{AbstractString}, Vector{Function}}("avg_fit" => Statistics.mean, "median_fit" => Statistics.median, "best_fit" => minimum), NamedTuple{(:avg_fit, :median_fit, :best_fit)}[])

Now we can use the built-in PSO algorithm:

EvoLP.PSO — Function

PSO(f, population, k_max; w=1, c1=1, c2=1)
PSO(logger::Logbook, f, population, k_max; w=1, c1=1, c2=1)

Arguments

f::Function: Objective function to minimise.

population::Vector{Particle}: a list of Particle individuals.

k_max::Integer: number of iterations.

Keywords

w: inertia weight. Optional, by default 1.

c1: cognitive coefficient (own's position). Optional, by default 1.

c2: social coefficient (others' position). Optional, by default 1.

Returns a Result.

Let's use the default parameters, and 30 iterations:

result = PSO(logbook, michalewicz, population, 30);

Result(-1.8012084267813209, [2.205109164422867, 1.57142448681964], Particle[Particle([0.6002390584670829, 1.7028640446466636], [-2.484353812982678, -0.06089319074782626], -0.460208354766287, [2.234980298639271, 1.5327931226043834], -1.7289677579182903), Particle([4.836247618106196, 1.4923187222834313], [-0.6539120149342456, 0.06374465661730949], -0.6105499534791514, [2.443706708415874, 1.5852872745354436], -1.2032828062174175), Particle([1.7986080095392907, 1.5433005171335894], [0.5371048305076176, 0.00845741735383137], -1.0149968038750694, [2.2607444053252905, 1.5673754161115518], -1.7469195586031856), Particle([2.7705777209566538, 1.571045345908194], [-0.8377919532625013, -0.0010050622391402874], -1.0000501604346417, [2.187884457230974, 1.5731587689916422], -1.7974418470060898), Particle([1.719452041070065, 1.3827305557929002], [0.05831777438408253, -0.22189137439733692], -0.2878275261124974, [2.2579378210619767, 1.5614016882696216], -1.748852919791968), Particle([2.2406710960833056, 1.4238074331933863], [2.3648119503402487, -0.02666481227909348], -1.2244506053104802, [2.277400836045768, 1.5535318372641846], -1.701343441382237), Particle([1.9700051416665243, 1.548295649371367], [0.06435442292046809, 0.011244518403326543], -1.2725666071649058, [2.189963310179992, 1.5608560245059322], -1.794631609876227), Particle([1.4162794340493385, 1.8580455425509228], [0.9246223168148393, -0.16944663869808133], -0.014128401933492487, [2.151488015741321, 1.5397196694667001], -1.7226906188981972), Particle([3.3018543168127557, 1.1912846221821483], [-1.876751707796851, 0.015253402801437277], -0.007423366006309035, [1.9071536439643955, 1.540672044575168], -1.127493893204107), Particle([2.9121133515841238, 2.3233004235052137], [-0.3342848068198338, 2.1384251766534663], -9.657992977635249e-9, [2.2014974089050794, 1.584883122086723], -1.7931948490117178)  …  Particle([2.153510476244535, 1.5858770177609207], [-0.013670181002566225, 0.507066253193933], -1.7541362748177838, [2.238771318304601, 1.5776679612238886], -1.7784201597107898), Particle([-0.23220261899398398, 0.387188061184546], [1.4804295341847757, 1.7358761264700506], -1.4398154962341737e-21, [1.84557088304814, 1.5894067562329073], -1.067520872149812), Particle([1.506959540497109, 0.2937977036607573], [0.5391262242315887, -0.24748110451400462], -0.00025711000850675217, [0.13466589212879154, 1.561653064872346], -0.9966391425879653), Particle([2.2036364377795103, 1.5727602014639255], [0.0009804839108448822, 0.00011450682255586772], -1.8011383428066545, [2.205109164422867, 1.57142448681964], -1.8012084267813209), Particle([-0.6882112555957813, -3.1715373109819147], [4.640567799002869, 7.7741843447400045], 2.165672168954497e-17, [2.306686947120561, 1.4063148409842705], -1.0044297875571326), Particle([2.4565961684589475, 1.3569818807552803], [0.0007413089499031522, -0.18476829281374169], -0.3719551598512881, [2.176028318868664, 1.623810070237835], -1.6785866854514344), Particle([2.2090360535726625, 1.5767439379253019], [-9.132757933295746e-5, 0.34396740940675596], -1.7992541555925063, [2.2090360535726625, 1.5767439379253019], -1.7992541555925063), Particle([2.8172359317199547, 1.5187262894340465], [0.20785991962349598, -0.008145760399953024], -0.8990465482541921, [2.230342028131265, 1.546234537421286], -1.7652353286240494), Particle([2.3787055874022256, 1.5945326486844742], [-0.03733808106075107, 0.028609010367763314], -1.3817889651249078, [2.2711613353439426, 1.535497330651596], -1.6786928298789943), Particle([2.0718279519618625, 1.5768741376976818], [0.25773315729009555, -0.014659648012348177], -1.5742189998354843, [2.227497051351536, 1.564803422313332], -1.7899804674334252)], 30, 1550, 0.30417799949645996)

The output was suppressed so that we can analyse each part of the result separately using the Result functions:

@show optimum(result)

-1.8012084267813209

@show optimizer(result)

2-element Vector{Float64}:
 2.205109164422867
 1.57142448681964

@show f_calls(result)

1550

We can also take a look at the logbook's records and see how the calculated statistics changed throughout the run:

for (i, I) in enumerate(logbook.records)
    print("it: $(i) with best_pos: $(I[3]) and avg_pos: $(I[1]) \n")
end

it: 1 with best_pos: -0.9660269227919166 and avg_pos: -0.0075582808175181345
it: 2 with best_pos: -0.9794114377120494 and avg_pos: -0.20300503450726065
it: 3 with best_pos: -1.0582436117258447 and avg_pos: -0.37423208936734964
it: 4 with best_pos: -1.188456899735513 and avg_pos: -0.3719007461097194
it: 5 with best_pos: -1.7791568176606942 and avg_pos: -0.5336394182167561
it: 6 with best_pos: -1.7226906188981972 and avg_pos: -0.6286405424121775
it: 7 with best_pos: -1.6994965493005854 and avg_pos: -0.6087654966157166
it: 8 with best_pos: -1.7974418470060898 and avg_pos: -0.7809637489713593
it: 9 with best_pos: -1.748852919791968 and avg_pos: -0.8112342427600603
it: 10 with best_pos: -1.7842887021263807 and avg_pos: -0.790180296430833
it: 11 with best_pos: -1.7907252497068264 and avg_pos: -0.8347781316182729
it: 12 with best_pos: -1.7895770952132557 and avg_pos: -0.8601146802224907
it: 13 with best_pos: -1.794652548523206 and avg_pos: -0.8433336422400548
it: 14 with best_pos: -1.7941114728329999 and avg_pos: -0.7765558656194087
it: 15 with best_pos: -1.7970090250002495 and avg_pos: -0.8347030415689519
it: 16 with best_pos: -1.7953820252918342 and avg_pos: -0.779334856289097
it: 17 with best_pos: -1.80097968438858 and avg_pos: -0.6816240443201734
it: 18 with best_pos: -1.7889881146870503 and avg_pos: -0.6968636891095334
it: 19 with best_pos: -1.7977362405675215 and avg_pos: -0.7885491671503049
it: 20 with best_pos: -1.8005192943710202 and avg_pos: -0.7235929089023462
it: 21 with best_pos: -1.7996004792318794 and avg_pos: -0.7024812138508402
it: 22 with best_pos: -1.8005740069812968 and avg_pos: -0.6876689591240711
it: 23 with best_pos: -1.8011350197832607 and avg_pos: -0.7620250970281185
it: 24 with best_pos: -1.8011166516863497 and avg_pos: -0.704588895763833
it: 25 with best_pos: -1.8010838647485024 and avg_pos: -0.6825427654291526
it: 26 with best_pos: -1.8010867455162902 and avg_pos: -0.714155312804364
it: 27 with best_pos: -1.8010613866676022 and avg_pos: -0.7477982780603933
it: 28 with best_pos: -1.8010728651434416 and avg_pos: -0.5813906640909421
it: 29 with best_pos: -1.8012084267813209 and avg_pos: -0.7337081197243474
it: 30 with best_pos: -1.8011383428066545 and avg_pos: -0.6650494389709172